+++ lib/widgetbook_app.directories.g.dart
@@ -11,45 +11,51 @@
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:autogram/ui/app_theme.dart' as _i3;
 import 'package:autogram/ui/assets.dart' as _i4;
-import 'package:autogram/ui/fragment/show_web_page_fragment.dart' as _i22;
-import 'package:autogram/ui/screens/about_screen.dart' as _i25;
-import 'package:autogram/ui/screens/main_menu_screen.dart' as _i26;
+import 'package:autogram/ui/fragment/document_validation_info_fragment.dart'
+    as _i26;
+import 'package:autogram/ui/fragment/preview_document_fragment.dart' as _i27;
+import 'package:autogram/ui/fragment/show_web_page_fragment.dart' as _i28;
+import 'package:autogram/ui/screens/about_screen.dart' as _i31;
+import 'package:autogram/ui/screens/main_menu_screen.dart' as _i32;
 import 'package:autogram/ui/screens/main_screen.dart' as _i2;
 import 'package:autogram/ui/screens/onboarding_accept_document_screen.dart'
-    as _i27;
-import 'package:autogram/ui/screens/onboarding_finished_screen.dart' as _i28;
+    as _i33;
+import 'package:autogram/ui/screens/onboarding_finished_screen.dart' as _i34;
 import 'package:autogram/ui/screens/onboarding_select_signing_certificate_screen.dart'
-    as _i29;
-import 'package:autogram/ui/screens/open_document_screen.dart' as _i30;
-import 'package:autogram/ui/screens/paired_device_list_screen.dart' as _i31;
+    as _i35;
+import 'package:autogram/ui/screens/open_document_screen.dart' as _i36;
+import 'package:autogram/ui/screens/paired_device_list_screen.dart' as _i37;
 import 'package:autogram/ui/screens/present_signed_document_screen.dart'
-    as _i32;
-import 'package:autogram/ui/screens/preview_document_screen.dart' as _i33;
-import 'package:autogram/ui/screens/qr_code_scanner_screen.dart' as _i20;
-import 'package:autogram/ui/screens/select_certificate_screen.dart' as _i34;
-import 'package:autogram/ui/screens/settings_screen.dart' as _i35;
-import 'package:autogram/ui/screens/show_document_screen.dart' as _i36;
-import 'package:autogram/ui/screens/sign_document_screen.dart' as _i37;
-import 'package:autogram/ui/screens/start_remote_document_signing_screen.dart'
     as _i38;
-import 'package:autogram/ui/widgets/app_version_text.dart' as _i9;
+import 'package:autogram/ui/screens/qr_code_scanner_screen.dart' as _i24;
+import 'package:autogram/ui/screens/select_certificate_screen.dart' as _i39;
+import 'package:autogram/ui/screens/settings_screen.dart' as _i40;
+import 'package:autogram/ui/screens/show_document_screen.dart' as _i41;
+import 'package:autogram/ui/screens/sign_document_screen.dart' as _i42;
+import 'package:autogram/ui/screens/start_remote_document_signing_screen.dart'
+    as _i43;
+import 'package:autogram/ui/widgets/app_version_text.dart' as _i10;
 import 'package:autogram/ui/widgets/autogram_logo.dart' as _i5;
 import 'package:autogram/ui/widgets/buttons.dart' as _i6;
-import 'package:autogram/ui/widgets/certificate_picker.dart' as _i23;
-import 'package:autogram/ui/widgets/close_button.dart' as _i7;
-import 'package:autogram/ui/widgets/dialogs.dart' as _i21;
-import 'package:autogram/ui/widgets/document_visualization.dart' as _i10;
-import 'package:autogram/ui/widgets/error_content.dart' as _i11;
-import 'package:autogram/ui/widgets/html_preview.dart' as _i12;
-import 'package:autogram/ui/widgets/loading_content.dart' as _i13;
-import 'package:autogram/ui/widgets/loading_indicator.dart' as _i8;
-import 'package:autogram/ui/widgets/markdown_text.dart' as _i14;
-import 'package:autogram/ui/widgets/option_picker.dart' as _i15;
-import 'package:autogram/ui/widgets/preference_tile.dart' as _i16;
-import 'package:autogram/ui/widgets/result_view.dart' as _i17;
-import 'package:autogram/ui/widgets/retry_view.dart' as _i18;
-import 'package:autogram/ui/widgets/signature_type_picker.dart' as _i24;
-import 'package:autogram/ui/widgets/step_indicator.dart' as _i19;
+import 'package:autogram/ui/widgets/certificate_details.dart' as _i11;
+import 'package:autogram/ui/widgets/certificate_picker.dart' as _i29;
+import 'package:autogram/ui/widgets/chip.dart' as _i7;
+import 'package:autogram/ui/widgets/close_button.dart' as _i8;
+import 'package:autogram/ui/widgets/dialogs.dart' as _i25;
+import 'package:autogram/ui/widgets/document_signature_info.dart' as _i12;
+import 'package:autogram/ui/widgets/document_validation_strip.dart' as _i13;
+import 'package:autogram/ui/widgets/document_visualization.dart' as _i14;
+import 'package:autogram/ui/widgets/error_content.dart' as _i15;
+import 'package:autogram/ui/widgets/html_preview.dart' as _i16;
+import 'package:autogram/ui/widgets/loading_content.dart' as _i17;
+import 'package:autogram/ui/widgets/loading_indicator.dart' as _i9;
+import 'package:autogram/ui/widgets/markdown_text.dart' as _i18;
+import 'package:autogram/ui/widgets/option_picker.dart' as _i19;
+import 'package:autogram/ui/widgets/preference_tile.dart' as _i20;
+import 'package:autogram/ui/widgets/result_view.dart' as _i21;
+import 'package:autogram/ui/widgets/retry_view.dart' as _i22;
+import 'package:autogram/ui/widgets/signature_type_picker.dart' as _i30;
+import 'package:autogram/ui/widgets/step_indicator.dart' as _i23;
 import 'package:widgetbook/widgetbook.dart' as _i1;
 
 final directories = <_i1.WidgetbookNode>[
@@ -93,6 +99,10 @@ final directories = <_i1.WidgetbookNode>[
         name: 'Button',
         useCases: [
           _i1.WidgetbookUseCase(
+            name: 'elevated',
+            builder: _i6.previewElevatedButton,
+          ),
+          _i1.WidgetbookUseCase(
             name: 'filled',
             builder: _i6.previewFilledButton,
           ),
@@ -103,10 +113,17 @@ final directories = <_i1.WidgetbookNode>[
         ],
       ),
       _i1.WidgetbookLeafComponent(
+        name: 'Chip',
+        useCase: _i1.WidgetbookUseCase(
+          name: '',
+          builder: _i7.previewChip,
+        ),
+      ),
+      _i1.WidgetbookLeafComponent(
         name: 'CloseButton',
         useCase: _i1.WidgetbookUseCase(
           name: 'CloseButton',
-          builder: _i7.previewCloseButton,
+          builder: _i8.previewCloseButton,
         ),
       ),
       _i1.WidgetbookLeafComponent(
@@ -120,7 +137,7 @@ final directories = <_i1.WidgetbookNode>[
         name: 'LoadingIndicator',
         useCase: _i1.WidgetbookUseCase(
           name: 'LoadingIndicator',
-          builder: _i8.previewLoadingIndicator,
+          builder: _i9.previewLoadingIndicator,
         ),
       ),
       _i1.WidgetbookLeafComponent(
@@ -153,56 +170,83 @@ final directories = <_i1.WidgetbookNode>[
         name: 'AppVersionText',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i9.previewAppVersionText,
+          builder: _i10.previewAppVersionText,
         ),
+      ),
+      _i1.WidgetbookLeafComponent(
+        name: 'CertificateDetails',
+        useCase: _i1.WidgetbookUseCase(
+          name: 'CertificateDetails',
+          builder: _i11.previewCertificateDetails,
+        ),
+      ),
+      _i1.WidgetbookLeafComponent(
+        name: 'DocumentSignatureInfo',
+        useCase: _i1.WidgetbookUseCase(
+          name: '',
+          builder: _i12.previewDocumentSignatureInfo,
+        ),
+      ),
+      _i1.WidgetbookComponent(
+        name: 'DocumentValidationStrip',
+        useCases: [
+          _i1.WidgetbookUseCase(
+            name: 'loading',
+            builder: _i13.previewLoadingDocumentValidationStrip,
+          ),
+          _i1.WidgetbookUseCase(
+            name: 'signatures',
+            builder: _i13.previewOtherDocumentValidationStrip,
+          ),
+        ],
       ),
       _i1.WidgetbookLeafComponent(
         name: 'DocumentVisualization',
         useCase: _i1.WidgetbookUseCase(
           name: 'DocumentVisualization',
-          builder: _i10.previewDocumentVisualization,
+          builder: _i14.previewDocumentVisualization,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'ErrorContent',
         useCase: _i1.WidgetbookUseCase(
           name: 'ErrorContent',
-          builder: _i11.previewErrorContent,
+          builder: _i15.previewErrorContent,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'HtmlPreview',
         useCase: _i1.WidgetbookUseCase(
           name: 'HtmlPreview',
-          builder: _i12.previewHtmlPreview,
+          builder: _i16.previewHtmlPreview,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'LoadingContent',
         useCase: _i1.WidgetbookUseCase(
           name: 'LoadingContent',
-          builder: _i13.previewLoadingContent,
+          builder: _i17.previewLoadingContent,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'MarkdownText',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i14.previewMarkdownText,
+          builder: _i18.previewMarkdownText,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'OptionPicker',
         useCase: _i1.WidgetbookUseCase(
           name: 'OptionPicker',
-          builder: _i15.previewOptionPicker,
+          builder: _i19.previewOptionPicker,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'PreferenceTile',
         useCase: _i1.WidgetbookUseCase(
           name: 'PreferenceTile',
-          builder: _i16.previewPreferenceTile,
+          builder: _i20.previewPreferenceTile,
         ),
       ),
       _i1.WidgetbookComponent(
@@ -210,19 +254,19 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'custom',
-            builder: _i17.previewCustomResultView,
+            builder: _i21.previewCustomResultView,
           ),
           _i1.WidgetbookUseCase(
             name: 'error',
-            builder: _i17.previewErrorResultView,
+            builder: _i21.previewErrorResultView,
           ),
           _i1.WidgetbookUseCase(
             name: 'info',
-            builder: _i17.previewInfoResultView,
+            builder: _i21.previewInfoResultView,
           ),
           _i1.WidgetbookUseCase(
             name: 'success',
-            builder: _i17.previewSuccessResultView,
+            builder: _i21.previewSuccessResultView,
           ),
         ],
       ),
@@ -230,21 +274,28 @@ final directories = <_i1.WidgetbookNode>[
         name: 'RetryView',
         useCase: _i1.WidgetbookUseCase(
           name: 'RetryView',
-          builder: _i18.previewRetryView,
+          builder: _i22.previewRetryView,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'StepIndicator',
         useCase: _i1.WidgetbookUseCase(
           name: 'StepIndicator',
-          builder: _i19.previewStepIndicator,
+          builder: _i23.previewStepIndicator,
+        ),
+      ),
+      _i1.WidgetbookLeafComponent(
+        name: '_InfoPanel',
+        useCase: _i1.WidgetbookUseCase(
+          name: '',
+          builder: _i24.previewInfoPanel,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: '_ViewFinder',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i20.previewViewFinder,
+          builder: _i24.previewViewFinder,
         ),
       ),
     ],
@@ -256,7 +307,7 @@ final directories = <_i1.WidgetbookNode>[
         name: 'BottomSheet',
         useCase: _i1.WidgetbookUseCase(
           name: 'NotificationsPermissionRationale',
-          builder: _i21.previewNotificationsPermissionRationaleModal,
+          builder: _i25.previewNotificationsPermissionRationaleModal,
         ),
       )
     ],
@@ -265,12 +316,36 @@ final directories = <_i1.WidgetbookNode>[
     name: 'Fragments',
     children: [
       _i1.WidgetbookLeafComponent(
+        name: 'DocumentValidationInfoFragment',
+        useCase: _i1.WidgetbookUseCase(
+          name: '',
+          builder: _i26.previewDocumentValidationInfoFragment,
+        ),
+      ),
+      _i1.WidgetbookComponent(
+        name: 'PreviewDocumentFragment',
+        useCases: [
+          _i1.WidgetbookUseCase(
+            name: 'error',
+            builder: _i27.previewErrorPreviewDocumentScreen,
+          ),
+          _i1.WidgetbookUseCase(
+            name: 'loading',
+            builder: _i27.previewLoadingPreviewDocumentFragment,
+          ),
+          _i1.WidgetbookUseCase(
+            name: 'success',
+            builder: _i27.previewSuccessPreviewDocumentScreen,
+          ),
+        ],
+      ),
+      _i1.WidgetbookLeafComponent(
         name: 'ShowWebPageFragment',
         useCase: _i1.WidgetbookUseCase(
           name: 'ShowWebPageFragment',
-          builder: _i22.previewShowWebPageFragment,
+          builder: _i28.previewShowWebPageFragment,
         ),
-      )
+      ),
     ],
   ),
   _i1.WidgetbookCategory(
@@ -280,14 +355,14 @@ final directories = <_i1.WidgetbookNode>[
         name: 'CertificatePicker',
         useCase: _i1.WidgetbookUseCase(
           name: 'CertificatePicker',
-          builder: _i23.previewCertificatePicker,
+          builder: _i29.previewCertificatePicker,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'SignatureTypePicker',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i24.previewSignatureTypePicker,
+          builder: _i30.previewSignatureTypePicker,
         ),
       ),
     ],
@@ -299,14 +374,14 @@ final directories = <_i1.WidgetbookNode>[
         name: 'AboutScreen',
         useCase: _i1.WidgetbookUseCase(
           name: 'AboutScreen',
-          builder: _i25.previewAboutScreen,
+          builder: _i31.previewAboutScreen,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'MainMenuScreen',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i26.previewMainMenuScreen,
+          builder: _i32.previewMainMenuScreen,
         ),
       ),
       _i1.WidgetbookLeafComponent(
@@ -320,14 +395,14 @@ final directories = <_i1.WidgetbookNode>[
         name: 'OnboardingAcceptDocumentScreen',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i27.previewOnboardingAcceptDocumentScreen,
+          builder: _i33.previewOnboardingAcceptDocumentScreen,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'OnboardingFinishedScreen',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i28.previewOnboardingFinishedScreen,
+          builder: _i34.previewOnboardingFinishedScreen,
         ),
       ),
       _i1.WidgetbookComponent(
@@ -335,20 +410,20 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'canceled',
-            builder: _i29.previewCanceledOnboardingSelectSigningCertificateBody,
+            builder: _i35.previewCanceledOnboardingSelectSigningCertificateBody,
           ),
           _i1.WidgetbookUseCase(
             name: 'initial',
-            builder: _i29.previewInitialOnboardingSelectSigningCertificateBody,
+            builder: _i35.previewInitialOnboardingSelectSigningCertificateBody,
           ),
           _i1.WidgetbookUseCase(
             name: 'no certificate',
             builder:
-                _i29.previewNoCertificateOnboardingSelectSigningCertificateBody,
+                _i35.previewNoCertificateOnboardingSelectSigningCertificateBody,
           ),
           _i1.WidgetbookUseCase(
             name: 'success',
-            builder: _i29.previewSuccessOnboardingSelectSigningCertificateBody,
+            builder: _i35.previewSuccessOnboardingSelectSigningCertificateBody,
           ),
         ],
       ),
@@ -357,11 +432,11 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'error',
-            builder: _i30.previewErrorOpenDocumentScreen,
+            builder: _i36.previewErrorOpenDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'loading',
-            builder: _i30.previewLoadingOpenDocumentScreen,
+            builder: _i36.previewLoadingOpenDocumentScreen,
           ),
         ],
       ),
@@ -369,7 +444,7 @@ final directories = <_i1.WidgetbookNode>[
         name: 'PairedDeviceListScreen',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i31.previewPairedDeviceListScreen,
+          builder: _i37.previewPairedDeviceListScreen,
         ),
       ),
       _i1.WidgetbookComponent(
@@ -377,36 +452,19 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'error',
-            builder: _i32.previewErrorPresentSignedDocumentScreen,
+            builder: _i38.previewErrorPresentSignedDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'initial',
-            builder: _i32.previewInitialPresentSignedDocumentScreen,
+            builder: _i38.previewInitialPresentSignedDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'loading',
-            builder: _i32.previewLoadingPresentSignedDocumentScreen,
+            builder: _i38.previewLoadingPresentSignedDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'success',
-            builder: _i32.previewSuccessPresentSignedDocumentScreen,
-          ),
-        ],
-      ),
-      _i1.WidgetbookComponent(
-        name: 'PreviewDocumentScreen',
-        useCases: [
-          _i1.WidgetbookUseCase(
-            name: 'error',
-            builder: _i33.previewErrorPreviewDocumentScreen,
-          ),
-          _i1.WidgetbookUseCase(
-            name: 'loading',
-            builder: _i33.previewLoadingPreviewDocumentScreen,
-          ),
-          _i1.WidgetbookUseCase(
-            name: 'success',
-            builder: _i33.previewSuccessPreviewDocumentScreen,
+            builder: _i38.previewSuccessPresentSignedDocumentScreen,
           ),
         ],
       ),
@@ -414,7 +472,7 @@ final directories = <_i1.WidgetbookNode>[
         name: 'QRCodeScannerScreen',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i20.previewQRCodeScannerScreen,
+          builder: _i24.previewQRCodeScannerScreen,
         ),
       ),
       _i1.WidgetbookComponent(
@@ -422,23 +480,23 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'canceled',
-            builder: _i34.previewCanceledSelectCertificateScreen,
+            builder: _i39.previewCanceledSelectCertificateScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'error',
-            builder: _i34.previewErrorSelectCertificateScreen,
+            builder: _i39.previewErrorSelectCertificateScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'loading',
-            builder: _i34.previewLoadingSelectCertificateScreen,
+            builder: _i39.previewLoadingSelectCertificateScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'no certificate',
-            builder: _i34.previewNoCertificateSelectCertificateScreen,
+            builder: _i39.previewNoCertificateSelectCertificateScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'success',
-            builder: _i34.previewSuccessSelectCertificateScreen,
+            builder: _i39.previewSuccessSelectCertificateScreen,
           ),
         ],
       ),
@@ -446,14 +504,14 @@ final directories = <_i1.WidgetbookNode>[
         name: 'SettingsScreen',
         useCase: _i1.WidgetbookUseCase(
           name: 'SettingsScreen',
-          builder: _i35.previewSettingsScreen,
+          builder: _i40.previewSettingsScreen,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'ShowDocumentScreen',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i36.previewShowDocumentScreen,
+          builder: _i41.previewShowDocumentScreen,
         ),
       ),
       _i1.WidgetbookComponent(
@@ -461,15 +519,15 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'error',
-            builder: _i37.previewErrorSignDocumentScreen,
+            builder: _i42.previewErrorSignDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'loading',
-            builder: _i37.previewLoadingSignDocumentScreen,
+            builder: _i42.previewLoadingSignDocumentScreen,
           ),
           _i1.WidgetbookUseCase(
             name: 'success',
-            builder: _i37.previewSuccessSignDocumentScreen,
+            builder: _i42.previewSuccessSignDocumentScreen,
           ),
         ],
       ),
@@ -477,7 +535,7 @@ final directories = <_i1.WidgetbookNode>[
         name: 'StartRemoteDocumentSigningScreen',
         useCase: _i1.WidgetbookUseCase(
           name: '',
-          builder: _i38.previewStartRemoteDocumentSigningScreen,
+          builder: _i43.previewStartRemoteDocumentSigningScreen,
         ),
       ),
     ],